Working Syn

+++ model/tdma-sender.h
@@ -21,7 +21,7 @@ namespace ns3
     {
       uint16_t interval;
       uint8_t deviceType;
-      uint8_t mSyncByDevId;
+      uint16_t mSyncByDevId;
     };
     class TDMASender : public Application
     {
@@ -63,7 +63,7 @@ namespace ns3
       tm startTime;
       uint16_t m_interval;
       uint8_t deviceType;
-      uint8_t m_syncByDevId;
+      uint16_t m_syncByDevId;
 
 
       EventId m_sendEvent;
